Fix Issues 1 and 32
<rev> Fixing incorrect property data type metadata in the Python DAQmx
API. Raised by issue: https://github.com/ni/nidaqmx-python/issues/32
<priv> Turns out a number of UInt64 properties had incorrect metadata. I
went through all of the U64 properties in the API and found some others
that were wrong too.

<rev> Working around issue:
https://github.com/ni/nidaqmx-python/issues/1, where nidaqmx doesn't
work on Python 2.7.13 because of a ctypes bug.

<rev> Changing to install from pypi for nidaqmx_src since installing
from Build Services package dependencies seems broken in newer versions
of pythonservices. We have a requirements file to enforce versions, so
who cares.

<bud> http://review-board.natinst.com/r/202523/

+++ nidaqmx/_task_modules/in_stream.py
@@ -313,10 +313,10 @@ class InStream(object):
     @property
     def curr_read_pos(self):
         """
-        float: Indicates in samples per channel the current position in
+        long: Indicates in samples per channel the current position in
             the buffer.
         """
-        val = ctypes.c_double()
+        val = ctypes.c_ulonglong()
 
         cfunc = lib_importer.windll.DAQmxGetReadCurrReadPos
         if cfunc.argtypes is None:
@@ -324,7 +324,7 @@ class InStream(object):
                 if cfunc.argtypes is None:
                     cfunc.argtypes = [
                         lib_importer.task_handle,
-                        ctypes.POINTER(ctypes.c_double)]
+                        ctypes.POINTER(ctypes.c_ulonglong)]
 
         error_code = cfunc(
             self._handle, ctypes.byref(val))
@@ -1681,13 +1681,13 @@ class InStream(object):
     @property
     def total_samp_per_chan_acquired(self):
         """
-        float: Indicates the total number of samples acquired by each
+        long: Indicates the total number of samples acquired by each
             channel. NI-DAQmx returns a single value because this value
             is the same for all channels. For retriggered acquisitions,
             this value is the cumulative number of samples across all
             retriggered acquisitions.
         """
-        val = ctypes.c_double()
+        val = ctypes.c_ulonglong()
 
         cfunc = lib_importer.windll.DAQmxGetReadTotalSampPerChanAcquired
         if cfunc.argtypes is None:
@@ -1695,7 +1695,7 @@ class InStream(object):
                 if cfunc.argtypes is None:
                     cfunc.argtypes = [
                         lib_importer.task_handle,
-                        ctypes.POINTER(ctypes.c_double)]
+                        ctypes.POINTER(ctypes.c_ulonglong)]
 
         error_code = cfunc(
             self._handle, ctypes.byref(val))
